Move get_last_checkpoint_epoch
