+++ corebehrt/modules/trainer/setup.py
@@ -1,11 +1,15 @@
-import numpy as np
+import os
+import re
 from typing import List
+
+import numpy as np
 from torch.utils.data import WeightedRandomSampler
 
 
 def get_sampler(cfg, outcomes: List[int]):
     """Get sampler for training data.
     sample_weight: float. Adjusts the number of samples in the positive class.
+
     """
 
     def _inverse_sqrt(x):
@@ -21,3 +25,21 @@ def get_sampler(cfg, outcomes: List[int]):
         return sampler
     else:
         return None
+
+
+def get_last_checkpoint_epoch(checkpoint_folder: str) -> int:
+    """Returns the epoch of the last checkpoint."""
+    # Regular expression to match the pattern retry_XXX
+    pattern = re.compile(r"checkpoint_epoch(\d+)_end\.pt$")
+    max_epoch = None
+    for filename in os.listdir(checkpoint_folder):
+        match = pattern.match(filename)
+        if match:
+            epoch = int(match.group(1))
+            if max_epoch is None or epoch > max_epoch:
+                max_epoch = epoch
+    # Return the folder with the maximum retry number
+    if max_epoch is None:
+        raise ValueError("No checkpoint found in folder {}".format(checkpoint_folder))
+
+    return max_epoch
